createActivePokemon function to create ActivePokemon from a string.
Give triggers and Trainer Battle Triggers now create pokemon by using the createActivePokemon function instead of having duplicate code.
The parameters for the Give Trigger is now give instead of giveItem.
ActivePokemon can now be created with hold items. In the description of a pokemon for a trigger, add
Item: item name*
When specifying a custom move set for a pokemon, you must end the move set with the * character.