EVERYTHING YOU KNOW  ABOUT SENDING MESSAGES TO BATTLEVIEW IS WRONG
Well, mostly everything.  And the main difference is actually in how you send messages to the battle, not how the battle sends to the battle view; though they're both different all the same and everything seems to be super great. Actually, they're both really different:

Messages sent to the battle are a lot less specific now -- no longer send just an HP update, or just a Status update, but they'll now just send the activePokemon on over.  From there, the battle will individual send multiple messages to the battle view with mostly all types of updates.
Specifically, each call with an activepokemon will send an HP update, a max HP update, a status condition update, a type update, a name update, and a gender update.
Message sending for switching an activePokemon in remains exactly the same as before
Stat gains have a separate message send still, and also makes a call to the main activePokemon message
Image update still has a separate message (and still looks fantastic for animation)
Special update types are the same
Catching a Pokemon updates are the same
Gaining experience still has a separate message function (I bet you totes noticed level and exp weren't in the main one), level is always included here with a boolean as to whether it's a levelUp or not
Learning a move updates are the same

Fixed a bug with Trick-or-Treat/Forest's Curse which was saying things along the lines of "Bulbasaur changed to the Grass/Grass type!" -- it will just say Grass from now on.

Updated the item-switching messages for Switcheroo and Trick since they were really lame and non-descriptive before

Generalized the way Type-changing works -- and made Forecast and Multitype soooo much better and way less buggy

GUYS I FIXED ILLUSION (LIKE THE BUGGIEST ABILITY EVER BEFORE) AND IT'S SEEMS REALLY GREAT RIGHT NOW I LOVE IT SO MUCH (EXCEPT FOR THAT ONE PART) BUT IT SEEMS REALLY NOT BUGGY AND IT'S SUPER WAY BETTER THAN BEFORE
Added a NameChanger class to fix all the weird annoying bugs -- the one thing I don't like is that I made the Invoke method in getName() only look at the ability because I don't want to pass the battle into that method so it's fine because Illusion is so far the only NameChanger, but if a non-ability came out that was, I would just need to be careful to update it.
Exciting things: IT NO LONGER SAYS "GO! BULBASAUR!" WHEN THE ILLUSION POKEMON IS A RAPIDASH (OR ANYTHING) WHICH IS THE BEST BUG FIX EVER BECAUSE THAT BUG COMPLETELY DEFEATED THE PURPOSE OF THE ABILITY AND MY SOLUTION IS NOT SUPER-HARDCODED AND IS VERY NICE AND CLEAN AND I DIDN'T THINK THAT WAS POSSIBLE FOR THIS PROBLEM (sorry for seeming annoyingly narcissistic -- this problem has been bothering me for over a year now and I'm just really excited to cross this off my todo-list)
It also no longer says the name of the Illusion Pokemon when you check your party and when it says "What will <name> do?" -- so it pretty much works as intended and that makes me smile
Transform now fails when used against a Pokemon under the effects of Illusion
Also: Gen  6 addition -- now goes to the last conscious Pokemon in the party (instead of failing if the last Pokemon is not conscious)

The additional items parameter for the getEffectsList function in battle now adds these parameters first in case you need a priority system for your invokes -- like the changeType effect

Fixed a bug with Powder moves failing on grass types -- don't worry about it